refactor: move profile repository and user dataclass
  * move the repository interface and implementation to model\profile
