+++ app/src/main/java/com/android/sample/model/profile/ProfilesRepository.kt
@@ -1,4 +1,4 @@
-package com.android.sample.model
+package com.android.sample.model.profile
 
 interface ProfilesRepository {
   fun getUser(userId: String, onSuccess: (User?) -> Unit, onFailure: (Exception) -> Unit)
