+++ app/src/main/java/com/android/sample/model/profile/ProfilesRepositoryFirestore.kt
@@ -1,4 +1,4 @@
-package com.android.sample.model
+package com.android.sample.model.profile
 
 import android.util.Log
 import com.google.firebase.firestore.DocumentSnapshot
@@ -7,7 +7,7 @@ import com.google.firebase.firestore.FirebaseFirestore
 class ProfilesRepositoryFirestore(private val db: FirebaseFirestore) : ProfilesRepository {
 
   override fun getUser(userId: String, onSuccess: (User?) -> Unit, onFailure: (Exception) -> Unit) {
-    db.collection("profiles").document(userId).get().addOnCompleteListener { task ->
+    db.collection("profiles").document("profiles-$userId").get().addOnCompleteListener { task ->
       if (task.isSuccessful) {
         val document = task.result
         if (document != null) {
